* testing:
  0.5.0 deployment

+++ aura/strike_tile/strike_tileController.js
@@ -1,6 +1,8 @@
 ({
     destroyTile: function (component, event, helper) {
-        if (component.get('v.destroyable')) {
+    	// Check to see if this component is in Strike Fiddler so we stop user from removing it
+    	let isNotStrikeDemo = component.getElement().parentElement.classList[0] !== 'strikeDemoOnly';
+        if (component.get('v.destroyable') && isNotStrikeDemo) {
             helper.notifyParent(component);
             helper.destroyComponent(component);
         }
